requirements.txt
azure-functions
azure-ai-openai